Solve part4

+++ src/main/scala/file_handler/FileHandler.scala
@@ -1,11 +1,10 @@
 package file_handler
 
 import org.apache.spark.sql.{DataFrame, SaveMode, SparkSession}
-
 import java.io.File
 import java.nio.file.{Files, StandardCopyOption}
 
-object CSVHandler {
+object FileHandler {
   private val fileSeparator: String = java.io.File.separator
   private val outputPath: String = f"data$fileSeparator%soutput$fileSeparator"
 
@@ -22,23 +21,26 @@ object CSVHandler {
   }
 
   /**
-   * Workaround since I can't figure out how to correctly name the created CSV files.
+   * Workaround since I can't figure out how to correctly name the created files
    *
-   * Loop through the output path and remove the .crc & SUCCESS  files, then rename the CSV file to
+   * Loop through the output path and remove the .crc & SUCCESS  files, then rename the created file to
    * the given name and move it to data/output.
    *
-   * On exit, remove the directory created by the saveToCsv() method.
+   * On exit, remove the directory created by the save methods.
    * */
   private def handleCreatedOutput(filename: String): Unit = {
-    val directory = new File(outputPath)
+    val directory = new File(outputPath+f"dir_$filename")
     if (directory.isDirectory) {
       val files = directory.listFiles()
       for (file <- files) {
         //remove .crc and SUCCESS files
-        if (file.getName.contains(".crc") || !file.getName.contains(".csv")) {
+        if (file.getName.contains("crc") || (!file.getName.contains(".csv") && !file.getName.contains("parquet"))) {
+          //println("Deleting = " + file.getName)
           file.delete()
         } else {
-          val dest = new File(outputPath + fileSeparator + filename)
+          val dest = new File(outputPath + filename)
+
+          //println(f"Moving ${file.toPath} -> ${dest.toPath}")
           Files.move(file.toPath, dest.toPath, StandardCopyOption.ATOMIC_MOVE)
         }
       }
@@ -55,8 +57,7 @@ object CSVHandler {
    * */
 
   def saveToCsv(df: DataFrame, fileName: String, delimiter: String = ","): Unit = {
-    println(f"Writing $df to a file = $outputPath" + fileName)
-
+    println(f"Writing $df to a CSV file = $outputPath" + fileName)
     df
       .coalesce(1)
       .write.mode(SaveMode.Overwrite)
@@ -64,8 +65,22 @@ object CSVHandler {
       .option("delimiter", delimiter)
       .option("escape", "")
       .mode("overwrite")
-      .csv(outputPath)
+      .csv(outputPath+f"dir_$fileName")
 
     handleCreatedOutput(fileName)
+  }
+
+  def saveAsParquet(df: DataFrame, fileName: String): Unit = {
+    println(f"Writing $df to a parquet file = $outputPath" + fileName)
+
+    df
+      .coalesce(1)
+      .write
+      .format("parquet")
+      .option("compression", "gzip")
+      .save(outputPath+f"dir_$fileName")
+
+    handleCreatedOutput(fileName)
+
   }
 }
